requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via pip-tools
django-taggit==0.22.2
django==2.0
first==2.0.1              # via pip-tools
mistune==0.8.3
olefile==0.44             # via pillow
pillow==4.3.0
pip-tools==1.11.0
psycopg2==2.7.3.2
pyprind==2.11.1
pytz==2017.3              # via django
six==1.11.0               # via pip-tools
